% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depths.R
\name{depth.R1}
\alias{depth.R1}
\title{Computes the amplitude and phase depths of the R1 trajectories in \code{fmat}}
\usage{
depth.R1(fmat)
}
\arguments{
\item{fmat}{Matrix of trajectories in R1. Each column is a function.}
}
\value{
amplitude and phase depths in a list
}
\description{
Computes the amplitude and phase depths of the R1 trajectories in \code{fmat}
}
