% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depths.R
\name{depth.R2}
\alias{depth.R2}
\title{Computes the amplitude and phase depths of the R2 trajectories in \code{tmat}}
\usage{
depth.R2(tmat)
}
\arguments{
\item{tmat}{Array of trajectories in R2. First axis is the X, Y coordinates. Second axis
is the trajectory value, third axis indexes the trajetories.}
}
\value{
amplitude and phase depths in a list
}
\description{
Computes the amplitude and phase depths of the R2 trajectories in \code{tmat}
}
