% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depths.R
\name{depth.S2}
\alias{depth.S2}
\title{Computes the amplitude and phase depths of the S2 trajectories in \code{tmat}}
\usage{
depth.S2(tmat)
}
\arguments{
\item{tmat}{Array of trajectories in S2. First axis is the X, Y, Z coordinates. Second axis
is the trajectory value, third axis indexes the trajetories.}
}
\value{
amplitude and phase depths in a list
}
\description{
Computes the amplitude and phase depths of the S2 trajectories in \code{tmat}
}
\note{
3D euclidean coordinates are automatically converted to spherical coordinates.
}
