% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{elastic_outliers}
\alias{elastic_outliers}
\title{Computes the amplitude and phase depths of the S2 trajectories in \code{tmat}}
\usage{
elastic_outliers(elastic_depths, ka = 1.8, thresh = 0.5)
}
\arguments{
\item{elastic_depths}{elastic depths produced by one of the depth.R1, depth.R2, or 
depth.S2 functions. Both amplitude and phase depths need to be included.}

\item{ka}{Controls the outlier bound / whisker depth. Higher numbers make outlier
bound further away from the median. Defaults to 1.8}

\item{thresh}{Controls the amount of depth thresholding. Set from 0 to 1, where
0 means only the top 0% of functions can be outliers and 1 means the top 100% of 
functions can be outliers. Defaults to 0.5.}
}
\value{
amplitude and phase depths in a list
}
\description{
Computes the amplitude and phase depths of the S2 trajectories in \code{tmat}
}
