% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticdepth.R
\docType{package}
\name{elasticdepth}
\alias{elasticdepth}
\alias{elasticdepth-package}
\title{Elastic Depths}
\description{
Computes the elastic depths for R1, R2, and S2 trajectories and identifies outliers.
}
\author{
Trevor Harris
}
